let kernel use functions provided by libfenix
