+++ Cargo.toml
@@ -1,4 +1,13 @@
 [workspace]
-members = [
+members = [ 
     "kernel",
+    "libfenix"
 ]
+resolver = "2"
+
+[profile.release]
+panic = "abort"
+
+[profile.dev]
+panic = "abort"
+
