+++ kernel/Cargo.toml
@@ -4,14 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+libfenix = { path = "../libfenix" }
 
 [build-dependencies]
-
-[profile.release]
-panic = "abort"
-
-[profile.dev]
-panic = "abort"
 
 [[bin]]
 name = "kernel"
