+++ kernel/src/exceptions.rs
@@ -1,7 +1,4 @@
-use crate::peripherals::gpio::{
-    self,
-    pins::{GPIO1_21, GPIO1_22, GPIO1_23},
-};
+use libfenix::gpio::{self, pins::{GPIO1_21, GPIO1_22, GPIO1_23}};
 
 #[no_mangle]
 fn data_abort_handler() {
