+++ libfenix/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "libfenix"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
