+++ libfenix/src/sysclock.rs
@@ -1,0 +1,7 @@
+use crate::Syscall;
+
+pub fn millis() -> u32 {
+    let syscall = Syscall::Millis;
+    syscall.call().unwrap()
+}
+
